暂时删除 npm 自动打包机制

+++ platform-sms-admin/admin-web/src/main/java/com/courage/platform/sms/admin/loader/SmsAdapterLoader.java
@@ -19,7 +19,7 @@ public class SmsAdapterLoader {
 
     private static ExtensionLoader<OuterAdapter> EXTENSION_LOADER = ExtensionLoader.getExtensionLoader(OuterAdapter.class);
 
-    private static ConcurrentHashMap<Long, OuterAdapter> ADAPTER_MAP = new ConcurrentHashMap<>(16);
+    private static ConcurrentHashMap<Integer, OuterAdapter> ADAPTER_MAP = new ConcurrentHashMap<>(16);
 
     public void loadAdapter(SmsChannelConfig smsChannelConfig) {
         String adapterName = smsChannelConfig.getChannelType();
